Se crea el ViewModelTriangulo.kt de la actividad individual, Alexander Espinosa Montoya
